Snake Game V2
I've added many things to the game. Its still really buggy. I tried to a rotation feature depending on where the the user wants the snake to move. I still need to work on where each of the body parts are located in the snake. Future idea could be making the snake have some type of movement when turning. For example it would move the snake up, the snake head will go up then the second piece would go to where the head was originally located and this keeps on going for each part. This will help the snake have more of a snake movement.