SnakeGame V2
I've added many debugging features and outputted information which can be useful for the users to see there movements. I've also fixed the growth logic, It will add a shape behind the snake head and adds its to the end of the group. The moment of the snake still haven't been complete.